move tls_detect to tcp_accept_errors
