+++ linkerd/app/integration/src/tests/telemetry/mod.rs
@@ -3,7 +3,7 @@
 // particular appears to do nothing... T_T
 #![allow(unused_imports)]
 
-mod tls_detect;
+mod tcp_accept_errors;
 
 use crate::*;
 use std::io::Read;
